Add db fixtures and tests

+++ client/rust/shared/src/data/lessons/api.rs
@@ -1,24 +1,7 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::data::api::Api;
-
-// {
-//     "count": 11,
-//     "next": "http://127.0.0.1:8000/api/v1/lessons/?page=9&page_size=1",
-//     "previous": "http://127.0.0.1:8000/api/v1/lessons/?page=7&page_size=1",
-//     "results": [
-//         {
-//             "id": "4",
-//             "title": "aaa",
-//             "type": 0,
-//             "language1": 1,
-//             "language2": 3,
-//             "owner": "admin",
-//             "updated_at": 1720705788
-//         }
-//     ]
-// }
+//use crate::data::api::Api;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub(super) struct LessonsResponse {
@@ -36,5 +19,5 @@ pub(super) struct LessonResponse {
     pub language1: u8,
     pub language2: u8,
     pub owner: String,
-    pub updated_at: u128,
+    pub updated_at: i64,
 }
